[ci skip] Move service to root
